services/order-service/requirements.txt
Flask==3.0.0
prometheus-flask-exporter==0.23.0
psycopg2-binary==2.9.9
flask-cors==4.0.0
